% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{asd_file}
\alias{asd_file}
\title{Get location of a sample ASD file}
\usage{
asd_file()
}
\value{
a character vector storing the location of the sample ASD file
}
\description{
Utility function that retrieves the location of the sample ASD binary file on disc.
}
\examples{
fn <- asd_file()
fn
}
