% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/asdreader.r
\name{get_metadata}
\alias{get_metadata}
\title{Reads metadata header from ASD file}
\usage{
get_metadata(f)
}
\arguments{
\item{f}{character, path to ASD file}
}
\value{
a list storing the metadata information in the ASD header,
as documented here: \url{http://support.asdi.com/Document/Documents.aspx}
}
\description{
Reads metadata header from ASD file
}
\examples{
asd_fn <- asd_file()
md <- get_metadata(asd_fn)
names(md)
}
\author{
Pierre Roudier
}

