% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{soil.asd}
\alias{soil.asd}
\title{Sample ASD file containing a soil spectrum}
\format{A binary ASD file}
\description{
Sample ASD file containing a soil spectrum
}
\details{
The spectrum contained in this ASD file was collected
on a soil sample from New Zealand using the ASD FieldSpec 3 spectrometer.
The file version is ASD 8.0.
}
\examples{
# Access the location of the ASD file using the following command
fn <- asd_file()
fn
# This function is actually just a shorthand for
fn <- system.file("extdata", "soil.asd", package = "asdreader")
fn
}

