% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.r
\name{asd_file}
\alias{asd_file}
\title{Get location of a sample ASD file}
\usage{
asd_file()
}
\value{
a character vector storing the location of the sample ASD file
}
\description{
Get location of a sample ASD file
}
\examples{
fn <- asd_file()
fn
}

