% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/asdreader.r
\docType{package}
\name{asdreader}
\alias{asdreader}
\alias{asdreader-package}
\title{Reads ASD Binary Files in R.

\code{asdreader} is a simple driver to read spectroscopy data collected
using ASD FieldSpec spectrometer and stored in \code{*.asd} files.}
\description{
Reads ASD Binary Files in R.

\code{asdreader} is a simple driver to read spectroscopy data collected
using ASD FieldSpec spectrometer and stored in \code{*.asd} files.
}

