% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/asdreader.r
\name{get_spectra}
\alias{get_spectra}
\title{Reads reflectance from ASD file}
\usage{
get_spectra(f, type = "reflectance")
}
\arguments{
\item{f}{a vector of paths to ASD file(s)}

\item{type}{a character vector, which type of spectra to return. \code{"reflectance"}, \code{"raw"}, \code{"white_reference"} are currently supported}
}
\value{
a matrix of the spectrum contained in the ASD file(s)
}
\description{
Reads reflectance from ASD file
}
\examples{
asd_fn <- asd_file()
m <- get_spectra(asd_fn)
plot(m[1,], type = 'l')
}
\author{
Pierre Roudier
}

